now draws colors and ids correctly. squares can be dragged and dropped normally.
it appears that glew does not initialize correctly all the functions. some need to be initialized manually with glutGetProcAddress.
glutGetProcAddress is a freeglut extension to glut.

the purpose of the second agrument to glClearBuffer* is still not clear, but 0 seems to work.

--HG--
extra : convert_revision : svn%3Aff035674-5af9-2b42-bfa6-c382101e57d9/trunk%4012